part4 bloglist backend 4.21* 'Cambiar la solicitud DELETE para que un blog solo pueda ser eliminado por quien lo creó. Devolver un código y mensaje adecuado cuando el blog no contiene un token o no el del usuario que lo creó. Cambiar a String el los ID para saber si son los mismos y verificar que es el mismo usuario'

+++ part4/controllers/blogs.js
@@ -53,6 +53,15 @@ blogsRouter.delete('/:id', async (request, response) => {
         return response.status(404).send({ error: 'Blog not found' })
     }
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid'})
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(403).json({ error: 'only the creator can delete the blog' })
+    }
+
     await Blog.findByIdAndDelete(id)
     response.status(204).end()
 })
